requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.7.1=pypi_0
apex=0.1=pypi_0
atomicwrites=1.3.0=pypi_0
attrs=19.1.0=pypi_0
backports=1.0=py_2
backports-abc=0.5=pypi_0
backports-shutil-get-terminal-size=1.0.0=pypi_0
backports.functools_lru_cache=1.5=py_2
backports_abc=0.5=py27_0
blas=1.0=mkl
bleach=3.1.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2019.11.27=0
cairo=1.14.12=h8948797_3
certifi=2019.6.16=pypi_0
cffi=1.12.3=py27h2e261b9_0
configparser=3.7.4=pypi_0
contextlib2=0.6.0.post1=pypi_0
cuda100=1.0=0
cudatoolkit=10.0.130=0
cycler=0.10.0=py27_0
cython=0.29.14=pypi_0
cython-bbox=0.1.3=pypi_0
dbus=1.13.12=h746ee38_0
decorator=4.4.0=pypi_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
enum34=1.1.6=pypi_0
et-xmlfile=1.0.1=pypi_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
funcsigs=1.0.2=py27_0
functools32=3.2.3.post2=pypi_0
future=0.17.1=py27_0
futures=3.3.0=py27_0
glib=2.63.1=h5a9c865_0
graphite2=1.3.13=h23475e2_0
grpcio=1.22.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=pypi_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
huepy=0.9.8.1=pypi_0
icu=58.2=h9c2bf20_1
importlib-metadata=1.3.0=pypi_0
inotify=0.2.10=pypi_0
intel-openmp=2019.4=243
ipaddress=1.0.22=pypi_0
ipdb=0.12.2=pypi_0
ipykernel=4.10.0=pypi_0
ipython=5.8.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jasper=2.0.14=h07fcdf6_1
jdcal=1.4.1=pypi_0
jinja2=2.10.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=5.3.1=pypi_0
jupyter-console=5.2.0=pypi_0
jupyter-core=4.5.0=pypi_0
kiwisolver=1.1.0=py27he6710b0_0
lapsolver=1.0.2=dev_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.29.0=py27hd408876_0
markdown=3.1.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=2.2.3=py27hb69df0a_0
metric-learn=0.5.0=pypi_0
mistune=0.8.4=pypi_0
mkl=2019.4=243
mkl_fft=1.0.12=py27ha843d7b_0
mkl_random=1.0.2=py27hd81dba3_0
more-itertools=5.0.0=pypi_0
motmetrics=1.1.3=pypi_0
nbconvert=5.5.0=pypi_0
nbformat=4.4.0=pypi_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py27hfd86e86_0
nose=1.3.7=pypi_0
notebook=5.7.8=pypi_0
numba=0.45.1=py27h962f231_0
numpy=1.16.4=py27h7e9f1db_0
numpy-base=1.16.4=py27hde5b4d6_0
olefile=0.46=py27_0
opencv=3.4.2=py27h6fd60c2_1
openpyxl=2.6.4=pypi_0
openssl=1.1.1d=h7b6447c_3
packaging=19.2=pypi_0
pandas=0.24.2=pypi_0
pandocfilters=1.4.2=pypi_0
pathlib2=2.3.4=pypi_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=6.1.0=py27h34e0f95_0
pip=19.1.1=py27_0
pixman=0.38.0=h7b6447c_0
pluggy=0.13.1=pypi_0
prometheus-client=0.7.1=pypi_0
prompt-toolkit=1.0.16=pypi_0
protobuf=3.9.0=pypi_0
ptyprocess=0.6.0=pypi_0
py=1.8.0=pypi_0
py-opencv=3.4.2=py27hb342d67_1
pycparser=2.19=py27_0
pygments=2.4.2=pypi_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py27h05f1152_2
pyrsistent=0.15.4=pypi_0
pytest=4.6.7=pypi_0
python=2.7.16=h9bab390_0
python-dateutil=2.8.0=pypi_0
pytorch=1.2.0=py2.7_cuda10.0.130_cudnn7.6.2_0
pytorch-ignite=0.2.0=pypi_0
pytz=2019.3=py_0
pyzmq=18.0.2=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=4.5.2=pypi_0
readline=7.0=h7b6447c_5
scandir=1.10.0=pypi_0
scikit-learn=0.20.4=pypi_0
scipy=1.2.2=pypi_0
send2trash=1.5.0=pypi_0
setuptools=41.0.1=py27_0
simplegeneric=0.8.1=pypi_0
singledispatch=3.4.0.3=pypi_0
sip=4.19.8=py27hf484d3e_0
six=1.12.0=py27_0
sklearn=0.0=pypi_0
sqlite=3.29.0=h7b6447c_0
subprocess32=3.5.4=py27h7b6447c_0
tb-nightly=1.15.0a20190731=pypi_0
terminado=0.8.2=pypi_0
testpath=0.4.2=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.4.0=pypi_0
tornado=5.1.1=pypi_0
tqdm=4.32.2=pypi_0
traitlets=4.3.2=pypi_0
typing=3.7.4.1=pypi_0
wcwidth=0.1.7=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=0.15.5=pypi_0
wheel=0.33.4=py27_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.4=h14c3975_4
zipp=0.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
